refactor BasicAsset class
